requirements.txt
python-whois==0.8.0
dnspython==2.3.0
